Loading indicator in editor, removing repetitive code.
